Change type of VideoModule.resize args
darcs-hash:20060424150013-10d61-b8f907e9be23403ccf1b7f3ef73504cd5bf8a863.gz

+++ video.h
@@ -6,15 +6,13 @@
 #ifndef __VIDEO_H__
 #define __VIDEO_H__
 
-#include "types.h"
-
 typedef struct VideoModule VideoModule;
 struct VideoModule {
 	const char *name;
 	const char *help;
 	int (*init)(void);
 	void (*shutdown)(void);
-	void (*resize)(uint_least16_t w, uint_least16_t h);
+	void (*resize)(unsigned int w, unsigned int h);
 	void (*toggle_fullscreen)(void);
 	void (*vdg_reset)(void);
 	void (*vdg_vsync)(void);
